[refactor] Move docProcessor to tfidf package
