+++ internal/app/services/metricService/metricService_test.go
@@ -1,3 +1,5 @@
+//go:build unit || !integration
+
 package metricService_test
 
 import (
