+++ internal/app/services/tfidf/tfidfService_test.go
@@ -1,7 +1,9 @@
-package services_test
+//go:build unit || !integration
+
+package tfidf_test
 
 import (
-	"Stant/LestaGamesInternship/internal/app/services"
+	"Stant/LestaGamesInternship/internal/app/services/tfidf"
 	"maps"
 	"slices"
 	"strings"
@@ -12,7 +14,7 @@ func TestProcessReaderToTerms(t *testing.T) {
 	file := strings.NewReader("Random text for testing purposes only")
 
 	want := []string{"Random", "text", "for", "testing", "purposes", "only"}
-	got, err := services.ProcessReaderToTerms(file)
+	got, err := tfidf.ProcessReaderToTerms(file)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -25,7 +27,7 @@ func TestGetTermFrequency(t *testing.T) {
 	text := []string{"word", "hello", "hello", "world"}
 
 	want := map[string]uint64{"word": 1, "hello": 2, "world": 1}
-	got := services.GetTermFrequency(text)
+	got := tfidf.GetTermFrequency(text)
 
 	if !maps.Equal(got, want) {
 		t.Errorf("Wanted %v, got %v\n", want, got)
@@ -37,7 +39,7 @@ func TestCalculateIdf(t *testing.T) {
 	termFrequency := uint64(1)
 
 	want := 1.0
-	got := services.CalculateIdf(termsAmount, termFrequency)
+	got := tfidf.CalculateIdf(termsAmount, termFrequency)
 
 	if want != got {
 		t.Errorf("Wanted %f, got %f\n", want, got)
